Move component registry interfaces to runtime definitions from container definitions
