+++ packages/components/external-component-loader/src/urlRegistry.ts
@@ -5,8 +5,7 @@
 // tslint:disable: no-console
 import { IComponent, IComponentQueryableLegacy } from "@prague/component-core-interfaces";
 import { IFluidPackage, IPraguePackage } from "@prague/container-definitions";
-import { ComponentRegistryTypes, IComponentRegistry } from "@prague/container-runtime";
-import { ComponentFactoryTypes, IComponentFactory } from "@prague/runtime-definitions";
+import { ComponentFactoryTypes, ComponentRegistryTypes, IComponentFactory, IComponentRegistry } from "@prague/runtime-definitions";
 import { Deferred } from "@prague/utils";
 
 /**
