+++ packages/components/shared-text/src/index.ts
@@ -9,8 +9,8 @@ import "./publicpath";
 
 import { IRequest } from "@prague/component-core-interfaces";
 import { IContainerContext, IRuntime, IRuntimeFactory } from "@prague/container-definitions";
-import { ContainerRuntime, IComponentRegistry } from "@prague/container-runtime";
-import { IComponentContext, IComponentFactory } from "@prague/runtime-definitions";
+import { ContainerRuntime } from "@prague/container-runtime";
+import { IComponentContext, IComponentFactory, IComponentRegistry } from "@prague/runtime-definitions";
 // import { SharedString } from "@prague/sequence";
 import * as Snapshotter from "@prague/snapshotter";
 import * as sharedTextComponent from "./component";
