+++ packages/framework/aqueduct/src/helpers/simpleModuleInstantiationFactory.ts
@@ -7,8 +7,7 @@ import { IComponent } from "@prague/component-core-interfaces";
 import {
     IContainerContext, IRuntime, IRuntimeFactory,
 } from "@prague/container-definitions";
-import { ComponentRegistryTypes, IComponentRegistry } from "@prague/container-runtime";
-import { ComponentFactoryTypes, IComponentContext, IComponentFactory } from "@prague/runtime-definitions";
+import { ComponentFactoryTypes, ComponentRegistryTypes, IComponentContext, IComponentFactory, IComponentRegistry } from "@prague/runtime-definitions";
 import { SimpleContainerRuntimeFactory } from "./simpleContainerRuntimeFactory";
 
 /**
